fix percentage type issue

+++ cs.project07.pokemon/game/entites/PokedexEntry.cs
@@ -30,17 +30,17 @@ namespace cs.project07.pokemon
     {
 		protected int _pokedexID;
         protected string _name;
-        protected int _maxHealth;
+        protected float _maxHealth;
         protected Element _element;
         protected Attack[] _attacks;
 
 		public int PokedexID { get => _pokedexID; }
         public string Name { get => _name; }
-        public int MaxHealth { get => _maxHealth; }
+        public float MaxHealth { get => _maxHealth; }
         public Element Element { get => _element; }
         public Attack[] Attacks { get => _attacks; }
 
-        public PokedexEntry(int pokedexId, string name, Element element, int maxHealth, Attack[] attacks)
+        public PokedexEntry(int pokedexId, string name, Element element, float maxHealth, Attack[] attacks)
         {
 			_pokedexID = pokedexId;
 			_name = name;
